@@ -2,25 +2,67 @@
     "private": true,
     "scripts": {
         "dev": "npm run development",
-        "development": "mix",
+        "development": "npm run routes:generate && mix",
         "watch": "mix watch",
         "watch-poll": "mix watch -- --watch-options-poll=1000",
         "hot": "mix watch --hot",
-        "prod": "npm run production",
-        "production": "mix --production"
+        "prod": "npm run routes:generate && npm run production",
+        "production": "mix --production",
+        "setup": "php bin/composer.phar install && npm i",
+        "routes:generate": "php artisan ziggy:generate"
     },
     "devDependencies": {
-        "axios": "^0.21",
-        "laravel-mix": "^6.0.6",
+        "@types/dropzone": "^5.7.4",
+        "axios": "^0.24.0",
+        "clean-webpack-plugin": "^3.0.0",
+        "cross-env": "^7.0",
+        "deepmerge": "^4.2.2",
+        "laravel-mix": "^6.0.39",
         "lodash": "^4.17.19",
-        "postcss": "^8.1.14",
-        "vue-axios": "^3.2.2",
-        "vue-loader": "^15.9.6",
-        "vue-router": "^3.4.9",
-        "vue-template-compiler": "^2.6.12"
+        "pusher-js": "^7.5.0",
+        "resolve-url-loader": "^3.1.0",
+        "sass": "~1.32",
+        "sass-loader": "^9.0.3",
+        "sass-resources-loader": "^2.2.5",
+        "vue-cropperjs": "^4.2.0",
+        "vue-loader": "^15.9.8",
+        "vue-template-compiler": "^2.7.10",
+        "vuetifyjs-mix-extension": "^0.0.20",
+        "webpack": "^5.88.2"
     },
     "dependencies": {
-        "vue": "^2.6.12",
-        "webpack": "^5.88.2"
+        "@bugsnag/js": "^7.17.0",
+        "@bugsnag/plugin-vue": "^7.17.0",
+        "@fortawesome/fontawesome-free": "5.15.4",
+        "@typeform/embed": "^2.8.0",
+        "@voerro/vue-tagsinput": "^2.7.1",
+        "@vueuse/core": "^10.4.1",
+        "@vueuse/sound": "^2.0.1",
+        "@zip.js/zip.js": "^2.0.12",
+        "animate.css": "^4.1.1",
+        "audio-recorder-polyfill": "^0.4.1",
+        "compare-versions": "^6.0.0-rc.1",
+        "crypto-js": "^4.1.1",
+        "css-loader": "^6.7.1",
+        "dropzone": "^6.0.0-beta.2",
+        "extendable-media-recorder": "^7.1.4",
+        "extendable-media-recorder-wav-encoder": "^7.0.83",
+        "fix-webm-duration": "^1.0.5",
+        "howler": "^2.2.3",
+        "is-reachable": "^5.2.1",
+        "laravel-mix-clean": "^0.1.0",
+        "pinia": "^2.0.22",
+        "vue": "^2.7.0",
+        "vue-cli-plugin-vuetify": "^2.5.8",
+        "vue-cookie": "^1.1.4",
+        "vue-lazy-youtube-video": "^2.4.0",
+        "vue-moment": "^4.1.0",
+        "vue-router": "^3.5.2",
+        "vue-style-loader": "^4.1.3",
+        "vuescroll": "^4.17.3",
+        "vuetify": "2.6.13",
+        "vuetify-loader": "^1.7.3",
+        "wavesurfer.js": "^7.1.5",
+        "ziggy": "^2.4.0"
     }
 }
