@@ -5,35 +5,46 @@ const articleShow = () => import('./components/article/Show.vue' /* webpackChunk
 const CategoryCreate = () => import('./components/category/Add.vue' /* webpackChunkName: "resource/js/components/category/add" */)
 const CategoryEdit = () => import('./components/category/Edit.vue' /* webpackChunkName: "resource/js/components/category/edit" */)
 
-export const routes = [
-    {
-        name: 'home',
-        path: '/',
-        component: Welcome
-    },
-    {
-        name: 'articleList',
-        path: '/articles',
-        component: articleList
-    },
-    {
-        name: 'articleShow',
-        path: '/articles/:id',
-        component: articleShow
-    },
-    {
-        name: 'categoryList',
-        path: '/category',
-        component: CategoryList
-    },
-    {
-        name: 'categoryEdit',
-        path: '/category/:id/edit',
-        component: CategoryEdit
-    },
-    {
-        name: 'categoryAdd',
-        path: '/category/add',
-        component: CategoryCreate
-    }
-]
+import Vue from 'vue'
+import VueRouter from 'vue-router'
+
+Vue.use(VueRouter)
+
+const router = new VueRouter({
+    mode: 'history',
+
+    routes: [
+        {
+            name: 'home',
+            path: '/',
+            component: Welcome
+        },
+        {
+            name: 'articleList',
+            path: '/articles',
+            component: articleList
+        },
+        {
+            name: 'articleShow',
+            path: '/articles/:id',
+            component: articleShow
+        },
+        {
+            name: 'categoryList',
+            path: '/category',
+            component: CategoryList
+        },
+        {
+            name: 'categoryEdit',
+            path: '/category/:id/edit',
+            component: CategoryEdit
+        },
+        {
+            name: 'categoryAdd',
+            path: '/category/add',
+            component: CategoryCreate
+        }
+    ]
+})
+
+export default router
